refactor(codec): extract IR type declarations into CodecIr.kt
Move the ~19 core IR type declarations (FieldSpec + variants, the
dispatcher/codec shapes, ScalarKind, WireWidth, ConditionRef,
LengthSource, PayloadCodecSource, etc.) out of the 9.5k-line
CodecEmitter.kt into a sibling CodecIr.kt as top-level `internal` types.
All emit logic and extension functions (requireFixed, codecReceiver,
encode/decodeAccessor, dispatchValue*, batch/when helpers) stay in
CodecEmitter. RemainingComparisonOp moves too (a nested type the moved
ConditionRef references). CodecEmitter.kt 9494 -> 8387 lines.

Pure code-motion: CodecEmitter.kt diff is deletions only, generated code
is byte-identical (all 306 snapshot goldens unchanged), no logic changed.
First step toward landing the upcoming analyze/dispatch refactors in
focused files instead of growing the monolith.
